+++ src/depiction/calibration/methods/calibration_method_dummy.py
@@ -1,6 +1,6 @@
 from numpy.typing import NDArray
 from xarray import DataArray
-
+import numpy as np
 from depiction.calibration.methods.calibration_method import CalibrationMethod
 from depiction.image import MultiChannelImage
 
@@ -8,7 +8,9 @@ from depiction.image import MultiChannelImage
 class CalibrationMethodDummy(CalibrationMethod):
     """Returns the input data and creates some dummy coefficients to ensure compatibility."""
 
-    def extract_spectrum_features(self, peak_mz_arr: NDArray[float], peak_int_arr: NDArray[float]) -> DataArray:
+    def extract_spectrum_features(
+        self, peak_mz_arr: NDArray[np.float64], peak_int_arr: NDArray[np.float64]
+    ) -> DataArray:
         return DataArray([0], dims=["c"])
 
     def preprocess_image_features(self, all_features: MultiChannelImage) -> MultiChannelImage:
@@ -18,8 +20,8 @@ class CalibrationMethodDummy(CalibrationMethod):
         return features
 
     def apply_spectrum_model(
-        self, spectrum_mz_arr: NDArray[float], spectrum_int_arr: NDArray[float], model_coef: DataArray
-    ) -> tuple[NDArray[float], NDArray[float]]:
+        self, spectrum_mz_arr: NDArray[np.float64], spectrum_int_arr: NDArray[np.float64], model_coef: DataArray
+    ) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
         return spectrum_mz_arr, spectrum_int_arr
 
     def __repr__(self) -> str:
