+++ src/depiction/calibration/models/fit_model.py
@@ -1,9 +1,10 @@
 from typing import Union
 from numpy.typing import NDArray
 from depiction.calibration.models import LinearModel, PolynomialModel
+import numpy as np
 
 
-def fit_model(x: NDArray[float], y: NDArray[float], model_type: str) -> Union[LinearModel, PolynomialModel]:
+def fit_model(x: NDArray[np.float64], y: NDArray[np.float64], model_type: str) -> Union[LinearModel, PolynomialModel]:
     """Fits a model to the given data, with the particular model_type."""
     if len(x) < 3:
         # If there are not enough points, return a zero model.
